Updating world generators, making it possible to add world gens like ore, plants, and even structures.

+++ src/net/minecraft/world/biome/BiomeGenEnd.java
@@ -13,8 +13,8 @@ public class BiomeGenEnd extends BiomeGenBase
         this.spawnableWaterCreatureList.clear();
         this.spawnableCaveCreatureList.clear();
         this.spawnableMonsterList.add(new SpawnListEntry(EntityEnderman.class, 10, 4, 4));
-        this.topBlock = (byte)Block.dirt.blockID;
-        this.fillerBlock = (byte)Block.dirt.blockID;
+        this.topBlock = Block.dirt.blockID;
+        this.fillerBlock = Block.dirt.blockID;
         this.theBiomeDecorator = new BiomeEndDecorator(this);
     }
